split of captive portal related code
